requirements.txt
pyrtlsdr
numpy